only install can method in Class::Std classes
A can method needs to be installed to implement part of the behavior of
AUTOMETHOD. This should only have an impact on Class::Std classes, as it
is tied to the AUTOLOAD method that is also installed.

Modifying UNIVERSAL::can has an impact on the entire process. It slows
it down, as well as causing any bugs to impact the entire process. And
the implementation of the can wrapper does have a bug regarding SUPER::
methods.

Rather than modifying UNIVERSAL::can, just install a can method in
Class::Std classes. This limits the impact of any bugs or slowdowns.
It also ensures that the AUTOMETHOD feature is properly limited to
Class::Std classes.